l4util/{x86/amd64}/bitops: remove superfluous type casts

+++ l4util/include/ARCH-amd64/bitops_arch.h
@@ -96,7 +96,7 @@ l4util_test_bit(int b, const volatile l4_umword_t * dest)
      "memory", "cc"
      );
 
-  return (int)bit;
+  return bit;
 }
 
 /* bit test and set */
@@ -119,7 +119,7 @@ l4util_bts(int b, volatile l4_umword_t * dest)
      "memory", "cc"
      );
 
-  return (int)bit;
+  return bit;
 }
 
 /* bit test and reset */
@@ -142,7 +142,7 @@ l4util_btr(int b, volatile l4_umword_t * dest)
      "memory", "cc"
      );
 
-  return (int)bit;
+  return bit;
 }
 
 /* bit test and complement */
@@ -165,7 +165,7 @@ l4util_btc(int b, volatile l4_umword_t * dest)
      "memory", "cc"
      );
 
-  return (int)bit;
+  return bit;
 }
 
 /* bit scan reverse */
